==> analyses/01_data-prep-grid.Rmd <==
---
title: "Analyses of swisscom data"
subtitle: "Grid preparation"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, scales,  
       jsonlite, 
       DT,
       sf, tmap)

tmap_mode("view")

import::from("sjmisc", "frq")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->

# Ancillary data

## Municipality boundaries 

```{r eval=FALSE}
st_gg21 <- st_read("data-raw/swisstopo/swissboundaries3d_2021-07_2056_5728/SHAPEFILE_LV95_LN02/swissBOUNDARIES3D_1_3_TLM_HOHEITSGEBIET.shp",
                   as_tibble = TRUE) %>% 
  st_zm(drop = TRUE, what = "ZM") %>% 
  # move to LV03
  st_transform(21781) %>%
  # filter(! BFS_NUMMER %in% c(2391, 5391, 5394)) %>% 
  # exclude lakes
  # filter(OBJEKTART != "Kantonsgebiet") %>% 
  # exclude FL & enclaves
  filter(ICC == "CH") %>% 
  select(BFS_NUMMER, NAME, KANTONSNUM, GEM_TEIL) %>% 
  rename(GMDNR = BFS_NUMMER,
         GMDNAME = NAME,
         KTNR = KANTONSNUM) %>% 
  arrange(GMDNR)

write_rds(st_gg21, "data/swisstopo/st_gg21.Rds")
```

```{r echo=FALSE}
st_gg21 <- read_rds("data/swisstopo/st_gg21.Rds")
```

```{r eval=FALSE, include=FALSE}
st_cities_1 <- st_gg21 %>% 
  filter(GMDNR %in% c(261, 6621, 2701, 5586, 351))

st_cities_2 <- st_gg21 %>% 
  filter(GMDNR %in% c(230, 1061, 3203, 5192, 371))

# st_cities_1 %>% 
#   st_geometry() %>% 
#   plot()
```

Example of Bern  

```{r echo=FALSE}
st_gg21 %>% 
  filter(GMDNR %in% c(351)) %>% 
  qtm(borders = "red", fill = NULL)
```

## City quartiers

Data from [BfS](https://www.bfs.admin.ch/bfs/de/home/grundlagen/agvch/quartiergrenzen-schweizer-staedte.html).  

```{r eval=FALSE}
st_qg17 <- st_read("data-raw/BfS/ag-b-00.03-95-qg17/shp/quart17.shp",
                   as_tibble = TRUE) %>% 
  st_zm(drop = TRUE, what = "ZM") %>% 
  st_transform(21781) %>%
  select(-OBJECTID_1, -Flaeche) %>% 
  rename(GMDNR = GMDE,
         QNAME = NAME,
         QNR = NR) %>% 
  left_join(st_gg21 %>% 
              st_drop_geometry() %>% 
              select(GMDNR, GMDNAME)) %>% 
  relocate(geometry, .after = last_col())

write_rds(st_qg17, "data/swisstopo/st_qg17.Rds")
```

Example of Bern  

```{r echo=FALSE}
st_qg17 <- read_rds("data/swisstopo/st_qg17.Rds")

st_qg17 %>% 
  filter(GMDNR %in% c(351)) %>% 
  st_geometry() %>% 
  qtm(borders = "red", fill = NULL)
```

## STATPOP offset

Could be used to define `offset` for `st_make_grid`

```{r eval=FALSE}
statpop_20 <- read_delim("data-raw/BfS/ag-b-00.03-vz2020statpop/STATPOP2020.zip", 
                         delim = ";", escape_double = FALSE, trim_ws = TRUE)[] %>% 
  mutate_all(as.integer)

# statpop_20 %>% slice(1:10) %>% View()

write_rds(statpop_20, "data/BfS/statpop_20.Rds")
```

```{r}
offset_bfs <- read_rds("data/BfS/statpop_20.Rds") %>% 
  summarise(min_x = min(X_KOORD),
            min_y = min(Y_KOORD)) %>% 
  st_as_sf(coords = c("min_x", "min_y"), 
           crs = 21781,
           remove = FALSE)
```

## swisscom offset

Using two communities that are furthest away in southerly and westerly directions:  

**Chansy** to define `x`  
- Postal code `1284`  
- SFOS number	`6611`  

```{r echo=FALSE}
st_gg21 %>% 
  filter(GMDNR %in% c(6611)) %>% 
  st_geometry() %>% 
  qtm(borders = "red", fill = NULL)
```

**Chiasso** to define `y`  
- Postal code `6830`  
- SFOS number	`5250`  

```{r echo=FALSE}
st_gg21 %>% 
  filter(GMDNR %in% c(5250)) %>% 
  st_geometry() %>% 
  qtm(borders = "red", fill = NULL)
```

Tile definitions were pulled from API using `query_swisscom_heatmaps_api.py`.  

Points of grid were defined using lower left corner coordinates.  

`WGS84` coordinates were transformed to `LV03`. 

```{r}
chansy_grid <- fromJSON("data/swisscom/chansy_grid.json")
chansy_grid <- flatten(chansy_grid$tiles) %>% 
  as_tibble()

chansy_grid_sf <- chansy_grid %>% 
  st_as_sf(coords = c("ll.x", "ll.y"), 
           crs = 4326,
           remove = FALSE) %>% 
  st_transform(21781) %>% 
  mutate(x = st_coordinates(.)[, 1],
         y = st_coordinates(.)[, 2])
```

```{r include=FALSE}
chiasso_grid <- fromJSON("data/swisscom/chiasso_grid.json")
chiasso_grid <- flatten(chiasso_grid$tiles) %>% 
  as_tibble()

chiasso_grid_sf <- chiasso_grid %>% 
  st_as_sf(coords = c("ll.x", "ll.y"), 
           crs = 4326,
           remove = FALSE) %>% 
  st_transform(21781) %>% 
  mutate(x = st_coordinates(.)[, 1],
         y = st_coordinates(.)[, 2])
```

Chansy    

```{r echo=FALSE}
qtm(chansy_grid_sf)
```

Chiasso  

```{r echo=FALSE}
qtm(chiasso_grid_sf)
```

Coordinates of lower left corner were then obtained by getting minimum values  

```{r}
offset_swisscom <- 
  bind_rows(
    
    chansy_grid_sf %>% 
      st_drop_geometry() %>%
      summarise(min_x = min(x),
                min_y = min(y)) ,
    
    chiasso_grid_sf %>% 
      st_drop_geometry() %>%
      summarise(min_x = min(x),
                min_y = min(y)) 
  ) %>%
  summarise(min_x = min(min_x),
            min_y = min(min_y)) %>% 
  # needs rounding - transfrom error?
  mutate(min_y = round(min_y)) %>% 
  st_as_sf(coords = c("min_x", "min_y"), 
           crs = 21781,
           remove = FALSE)
```

Note small difference from BfS derived minimums (in black)!  

```{r echo=FALSE}
tm_shape(offset_bfs) +
  tm_dots(col = "black") +
  tm_shape(offset_swisscom) +
  tm_dots(col = "red")
```

```{r include=FALSE}
rm(chansy_grid, chiasso_grid,
   chansy_grid_sf, chiasso_grid_sf, offset_bfs); gc()
```

## Commuter flows 

### Data 

2018 [commuter flows](https://www.bfs.admin.ch/asset/de/ts-x-11.04.04.05-2018) from mobility microcensus.  
Testing only on areas of canton Bern.  

```{r}
commuters <- read_delim("data-raw/BfS/ts-x-11.04.04.05-2018.csv", 
                        delim = ";", escape_double = FALSE, 
                        col_types = cols(REF_YEAR = col_integer(),
                                         GEO_CANT_RESID = col_integer(),
                                         GEO_COMM_RESID = col_integer(), 
                                         GEO_CANT_WORK = col_integer(),
                                         GEO_COMM_WORK = col_integer(), 
                                         VALUE = col_integer()), 
                        trim_ws = TRUE) %>% 
  janitor::clean_names() %>% 
  filter(perspective == "R") %>% 
  select(-perspective, -ref_year) %>% 
  filter(geo_cant_resid == 2) %>% 
  select(-geo_cant_resid) %>% 
  arrange(geo_comm_resid, desc(value))
```

There are cross cantonal flows still here!  

```{r echo=FALSE}
frq(commuters, geo_cant_work)
```

### Aggregating by community of origin

```{r}
geo_comm_resid_agg <- 
  
  left_join(
    
    commuters %>% 
      group_by(geo_comm_resid) %>% 
      summarise(workers_all = sum(value)) %>% 
      ungroup() %>% 
      select(geo_comm_resid, workers_all),
    
    commuters %>% 
      filter(geo_comm_resid == geo_comm_work) %>% 
      rename(workers_stay = value) %>% 
      ungroup() %>% 
      select(geo_comm_resid, workers_stay)
    
  ) %>% 
  
  left_join(
    
    commuters %>% 
      filter(geo_comm_resid != geo_comm_work) %>% 
      group_by(geo_comm_resid) %>% 
      summarise(workers_leave = sum(value)) %>% 
      ungroup() %>% 
      select(geo_comm_resid, workers_leave) 
  ) %>% 
  mutate(percent_leave = (workers_leave / workers_all) * 100)
```

### Map of areas with high outflows 

```{r include=FALSE}
gem18 <- st_read("../../projects/ISPM_geo/data-raw/swisstopo/swissBOUNDARIES3D/BOUNDARIES_2018/DATEN/SHAPEFILE_LV95_LN02/swissBOUNDARIES3D_1_3_TLM_HOHEITSGEBIET.shp") %>% 
  st_zm(drop = TRUE) %>% 
  st_transform(2056) %>% 
  filter(ICC == "CH") %>% 
  filter(BFS_NUMMER < 9000) %>% 
  filter(OBJEKTART == "Gemeindegebiet") %>% 
  select(BFS_NUMMER, NAME, GEM_TEIL) %>% 
  rename(geo_comm_resid = BFS_NUMMER,
         geo_comm_name = NAME) %>% 
  arrange(geo_comm_resid, GEM_TEIL)

gem18 <- inner_join(gem18, geo_comm_resid_agg)

# anti_join(geo_comm_resid_agg, gem18)

plz_22_06 <- read_rds("../../projects/ISPM_geo/data/PLZ/plz_22_06.Rds") %>% 
  st_transform(2056) %>% 
  st_filter(gem18)
```

```{r echo=FALSE}
tm_shape(gem18) +
  tm_fill(col = "percent_leave", border.col = "grey40") +
  tm_text("geo_comm_name") + 
  tm_shape(plz_22_06) +
  tm_borders("blue", lwd = .5) +
  tm_text("PLZ", col = "blue")
```

### Bernese suburbs examples {.tabset}

#### Frauenkappelen 

```{r echo=FALSE}
commuters %>% 
  filter(geo_comm_resid == 663) %>% 
  datatable()
```

#### Bremgarten 

```{r echo=FALSE}
commuters %>% 
  filter(geo_comm_resid == 353) %>% 
  datatable()
```

#### Stettlen 

```{r echo=FALSE}
commuters %>% 
  filter(geo_comm_resid == 358) %>% 
  datatable()
```

<!-- ----------------------------------------------------- -->

# Grid

## Generate whole country

100m grid, with lower left corner defined using swisscom derived offset.  

Also, adding ID for each cell (based on row number).   

```{r eval=FALSE}
country <- st_gg21 %>% 
  st_make_grid(cellsize = 100, 
               offset = c(offset_swisscom$min_x, offset_swisscom$min_y),
               square = TRUE) %>% 
  st_sf() %>%
  st_cast("POLYGON") %>% 
  mutate(ID1 = row_number()) %>% 
  relocate(ID1)

# st_write(country, "data/grid/country.gpkg")
write_rds(country, "data/grid/country.Rds")
```

```{r echo=FALSE}
country <- read_rds("data/grid/country.Rds")
```

## Generate cities

Clipped / selected using swisstopo municipality data.  

Added second, city specific ID and area. 

```{r eval=FALSE, include=FALSE}
# tests

bern_clip <- country %>%
  st_intersection(st_gg21 %>% filter(GMDNR == 351))  %>%
  mutate(ID2 = row_number(),
         area = st_area(.)) %>%
  relocate(ID1, ID2, area)

bern1 <- country %>% 
  filter(st_intersects(geometry, 
                       st_gg21 %>% filter(GMDNR == 351), 
                       sparse = FALSE)) %>%
  mutate(ID2 = row_number()) %>%
  relocate(ID1, ID2)

bern2 <- country[st_gg21 %>% filter(GMDNR == 351), ] %>%
  mutate(ID2 = row_number()) %>% 
  relocate(ID1, ID2)

identical(bern1, bern2)
anti_join(bern1 %>% st_drop_geometry(), 
          bern2 %>% st_drop_geometry(), 
          by = "ID1")

rm(bern1, bern2); gc()
```

### Function with clipping 

```{r}
# why st_intersection 
# https://stackoverflow.com/questions/62442150/why-use-st-intersection-rather-than-st-intersects

city_grid_clip <- function(municipality){
  
  city <- country %>% 
    st_intersection(st_gg21 %>% 
                      filter(GMDNR == municipality))  %>%
    mutate(ID2 = row_number(),
           area = st_area(.)) %>% 
    relocate(ID1, ID2, area)
  
  return(city)
}
```

### Function without clipping 

```{r}
# solution from
# https://stackoverflow.com/questions/57014381/how-to-filter-an-r-simple-features-collection-using-sf-methods-like-st-intersect

city_grid <- function(municipality){
  
  city <- country %>% 
    filter(st_intersects(geometry, 
                         st_gg21 %>% filter(GMDNR == municipality), 
                         sparse = FALSE)) %>%
    mutate(ID2 = row_number()) %>% 
    relocate(ID1, ID2)
  
  return(city)
}
```

```{r}
# or even simpler, without dplyr
# could be generalized with arguments to provide full grid and GMDE shape

city_grid <- function(municipality){
  
  city <- country[st_gg21 %>% filter(GMDNR == municipality), ] %>%
    mutate(ID2 = row_number()) %>% 
    relocate(ID1, ID2)
  return(city)
  
}
```

### Zuirich

```{r}
# zurich_clip <- city_grid_clip(261)
zurich <- city_grid(261)
```

### Geneva

```{r}
# geneva_clip <- city_grid_clip(6621)
geneva <- city_grid(6621)
```

### Basel

```{r}
# basel_clip <- city_grid_clip(2701)
basel <- city_grid(2701)
```

### Lausanne

```{r}
# lausanne_clip <- city_grid_clip(5586)
lausanne <- city_grid(5586)
```

### Bern 

```{r}
bern_clip <- city_grid_clip(351)
bern <- city_grid(351)
```

```{r echo=FALSE}
# st_write(bern_clip, "data/grid/bern_clip.gpkg")
# st_write(bern, "data/grid/bern.gpkg")
write_rds(bern_clip, "data/grid/bern_clip.Rds")
write_rds(bern, "data/grid/bern.Rds")
```

### Offset 

Using two communities to get min x and y coordinates: *Chansy* & *Chiasso* 

```{r}
# chansy_clip <- city_grid_clip(6611)
# st_write(chansy_clip, "data/grid/chansy_clip.gpkg")
# write_rds(chansy_clip, "data/grid/chansy_clip.Rds")

chansy <- city_grid(6611)
# st_write(chansy, "data/grid/chansy.gpkg")
write_rds(chansy, "data/grid/chansy.Rds")

# chiasso_clip <- city_grid_clip(5250)
# st_write(chiasso_clip, "data/grid/chiasso_clip.gpkg")
# write_rds(chiasso_clip, "data/grid/chiasso_clip.Rds")

chiasso <- city_grid(5250)
# st_write(chiasso, "data/grid/chiasso.gpkg")
write_rds(chiasso, "data/grid/chiasso.Rds")
```

```{r eval=FALSE, include=FALSE}
chansy %>% 
  st_geometry() %>% 
  qtm(borders = "red", fill = NULL)

chiasso %>% 
  st_geometry() %>% 
  qtm(borders = "red", fill = NULL)
```

## Example 

Municipality Bern coverage:  

```{r eval=FALSE, include=FALSE}
bern_clip %>% 
  st_geometry() %>% 
  plot()

bern %>% 
  st_geometry() %>% 
  plot()
```

Clipped grid:  

```{r echo=FALSE}
bern_clip %>% 
  st_geometry() %>% 
  qtm(borders = "red", fill = NULL)
```

Vs. unclipped one:  

```{r echo=FALSE}
bern %>% 
  st_geometry() %>% 
  qtm(borders = "red", fill = NULL)
```

Note, when using clipped grid, there are few examples of grid cells with very small area.  
These could be excluded if no population was found there?

```{r echo=FALSE}
bern_clip %>% 
  st_drop_geometry() %>% 
  arrange(area) %>% 
  slice(1:10)
```

<!-- ----------------------------------------------------- -->

# Area estimates

## Bern - hourly

- Number of tiles for area of Bern  

```{r}
(tiles <- nrow(bern))
```


- Number of requests needed  

```{r}
(requests <- ceiling(tiles / 100))
```


- Price for complete grid  

```{r}
(price <- requests * 0.01)
```


- Price for 24h  

```{r}
(price_24h <- price * 24)
```


- Price for 4w hourly    

```{r}
(price_4w <- price_24h * 7 * 4)
```


- Price for year hourly   

```{r}
(price_year <- price_24h * 365)
```


- Price for 4w daily    

```{r}
(price_4w_d <- price * 7 * 4)
```


- Price for year daily   

```{r}
(price_year_d <- price * 365)
```

## 5 cities  - hourly

- Number of tiles for area of Bern  

```{r}
(tiles <- nrow(zurich) +
   nrow(geneva) +
   nrow(basel) +
   nrow(lausanne) +
   nrow(bern))
```


- Number of requests needed  

```{r}
(requests <- ceiling(tiles / 100))
```


- Price for complete grid  

```{r}
(price <- requests * 0.01)
```


- Price for 24h  

```{r}
(price_24h <- price * 24)
```


- Price for 4w hourly    

```{r}
(price_4w <- price_24h * 7 * 4)
```


- Price for year hourly   

```{r}
(price_year <- price_24h * 365)
```


- Price for 4w daily    

```{r}
(price_4w_d <- price * 7 * 4)
```


- Price for year daily   

```{r}
(price_year_d <- price * 365)
```

<!-- ----------------------------------------------------- -->

# Environment

```{r echo=FALSE, results='asis'}
report::report(sessionInfo())
```
