==> analyses/02_grid-further.Rmd <==
---
title: "Analyses of swisscom data"
subtitle: "Grid extras"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, scales,  
       jsonlite, 
       sf, tmap)

tmap_mode("view")

import::from("sjmisc", "frq")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->

# Swisscom grid coordinates & IDs

Tile definitions were pulled from API using `query_swisscom_heatmaps_api.py`.  

```{r}
read_tiles <- function(filename) {
  
  data <- jsonlite::fromJSON(filename)
  data <- jsonlite::flatten(data$tiles) %>% 
    dplyr::as_tibble()
  
  data$plz <- gsub("grid_|.json", "", filename)
  data$plz <- gsub("data/swisscom/", "", data$plz)
  
  return( data )
}

doFuture::registerDoFuture()
future::plan("multisession", workers = 8)

grid <- plyr::ldply(.data = fs::dir_ls("data/swisscom/", 
                                       regexp = "[0-9][.]json$"),
                    .fun = read_tiles,
                    .id = NULL,
                    .parallel = TRUE) %>% 
  as_tibble() %>% 
  distinct()
```

Focusing on test area of Bern city centre and selected suburbs, including postal codes:  

```{r echo=FALSE}
frq(grid, plz)
```

Points of grid were defined using lower left corner coordinates. They were also shifted by 50m east and north to better align with grids.    

```{r}
grid_sf <- grid %>% 
  st_as_sf(coords = c("ll.x", "ll.y"), 
           crs = 4326,
           remove = TRUE) %>% 
  st_transform(21781) %>% 
  mutate(x = st_coordinates(.)[, 1],
         y = st_coordinates(.)[, 2]) %>% 
  select(-ur.x, -ur.y)

# shifting by 50m to the centre
grid_sf_50 <- grid_sf %>% 
  st_drop_geometry() %>% 
  mutate(x = as.integer(as.integer(x) + 51), # why on earth 1?
         y = as.integer(as.integer(y) + 50)) %>% 
  st_as_sf(coords = c("x", "y"), 
           crs = 21781,
           remove = FALSE) 
```

```{r eval=FALSE, include=FALSE}
tm_shape(grid_sf) + 
  tm_dots(col = "grey50") + 
  tm_shape(grid_sf_50) + 
  tm_dots(col = "darkorchid")
```

<!-- ----------------------------------------------------- -->

# Grid derived with swisscom offset

swisscom points were linked to country grid derived in file `01.Rmd` providing access to crucial `tile ID` variable needed to link to the Heatmap API outputs.  

```{r eval=FALSE}
bern_plz <- 
  read_rds("data/grid/country.Rds") %>% 
  st_join(grid_sf_50,
          left = FALSE)

write_rds(bern_plz, "data/grid/bern_plz.Rds")
```

```{r include=FALSE}
bern_plz <- read_rds("data/grid/bern_plz.Rds")
```

```{r eval=FALSE, include=FALSE}
bern_plz_ct <- st_centroid(bern_plz)

tm_shape(grid_sf_50) + 
  tm_dots(col = "darkorchid") +
  tm_shape(bern_plz_ct) + 
  tm_dots(col = "grey50")
```

## Study area coverage  

```{r echo=FALSE}
tm_shape(bern_plz) + 
  tm_borders(col = "darkorchid") +
  tm_shape(grid_sf_50) + 
  tm_dots(col = "grey50")
```

## Duplicate cells

There are some cells in the grid that are duplicated because they overlap two (or more?) PLZs and were returned twice. 

```{r echo=FALSE}
frq(duplicated(bern_plz$tileId))
```

Example:  

```{r echo=FALSE}
tm_shape(bern_plz %>% filter(plz %in% c("3084", "3098"))) + 
  tm_fill(col = "plz", alpha = 0.5) + 
  tm_shape(bern_plz %>% filter(tileId == 31718585)) + 
  tm_borders(col = "red")
```

They do have unique ID so can easily be excluded in order to create correct visualizations (see [issue #8](https://github.com/RPanczak/FUN_swisscom/issues/8)). However analyses that would be based on PLZs, particularly aggregation of data would have to determine correct assignment of grid cells to PLZs. Perhaps by using (pop weighted?) centroid or sth similar?  

<!-- ----------------------------------------------------- -->

# Environment

```{r echo=FALSE, results='asis'}
report::report(sessionInfo())
```
