==> analyses/04_digital-api-eda.Rmd <==
---
title: "Analyses of swisscom data"
subtitle: "API data"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, scales,  
       lubridate, 
       sf, tmap)

tmap_mode("view")

import::from("sjmisc", "frq")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->

# Data 

## Grid 

Testing on several PLZs of Bern & surroundings.  

```{r}
bern_plz <- read_rds("data/grid/bern_plz.Rds") %>% 
  # duplicates are out for visualizations! 
  # fix if PLZ is needed more precisely!
  distinct(tileId, .keep_all = TRUE) %>% 
  rename(tile_id = tileId)

bern_plz_ct <- st_centroid(bern_plz)
```

Tiles per PLZ:  

```{r echo=FALSE}
frq(bern_plz$plz)
```

```{r eval=FALSE, include=FALSE}
View(st_drop_geometry(bern_plz))
```

```{r echo=FALSE}
# tm_shape(bern_plz) + 
#   tm_borders(col = "darkorchid") +
#   tm_shape(bern_plz_ct) + 
#   tm_dots(col = "grey50")

tm_shape(bern_plz) + 
  tm_fill(col = "plz", alpha = 0.5)
```

## Dwell density 

Data is from http://mip.swisscom.ch which swisscom describes as: 

> Our new API platform offering 3 endpoints focusing on *density*, *dwell times* and *origin destination*  

*Important note:* free data is limited to **2020-01-27** only!  

We are using [Heatmaps API](https://digital.swisscom.com/products/heatmaps/info) to retrieve daily and hourly dwell times for one postcode. Code to retrieve data, kindly provided by Yann Steimer from swisscom, is in `example_notebook_SC_heatmaps_API_UNIBE.ipynb`.   

### Daily dwell density

```{r}
read_day <- function(filename) {
  
  data <- readr::read_delim(filename, 
                            delim = ";", escape_double = FALSE, trim_ws = TRUE,
                            show_col_types = FALSE) %>% 
    dplyr::select(tile_id, time, score) %>% 
    dplyr::as_tibble()
  
  data$plz <- gsub("_day|.csv", "", filename)
  data$plz <- gsub("data/swisscom/", "", data$plz)
  
  return(data)
}

doFuture::registerDoFuture()
future::plan("multisession", workers = 8)

data_day <- plyr::ldply(.data = fs::dir_ls("data/swisscom/", 
                                           regexp = "[0-9]_day[.]csv$"),
                        .fun = read_day,
                        .id = NULL,
                        .parallel = TRUE) %>% 
  as_tibble() %>% 
  distinct(tile_id, time, .keep_all = TRUE)
```

### Hourly dwell density

```{r}
read_hour <- function(filename) {
  
  data <- readr::read_delim(filename, 
                            delim = ";", escape_double = FALSE, trim_ws = TRUE,
                            show_col_types = FALSE) %>% 
    dplyr::select(tile_id, time, score) %>% 
    dplyr::as_tibble()
  
  data$plz <- gsub("_hour|.csv", "", filename)
  data$plz <- gsub("data/swisscom/", "", data$plz)
  
  return(data)
}

data_hour <- plyr::ldply(.data = fs::dir_ls("data/swisscom/", 
                                            regexp = "[0-9]_hour[.]csv$"),
                         .fun = read_hour,
                         .id = NULL,
                         .parallel = TRUE) %>% 
  as_tibble() %>% 
  distinct(tile_id, time, .keep_all = TRUE)
```

```{r eval=FALSE, include=FALSE}
View(data_day)
View(data_hour)
```

<!-- ----------------------------------------------------- -->

# EDA

## Daily 

```{r}
bern_plz_day <- bern_plz %>% 
  left_join(data_day %>% select(-time))
```

```{r echo=FALSE}
bern_plz_day %>% 
  tm_shape() + 
  tm_polygons(col = "score", alpha = 0.5, palette = "-plasma",
              n = 7, style = "jenks",
              border.col = "white", border.alpha = 0.5)
```

## Hourly 

```{r}
bern_plz_hour <- bern_plz %>% 
  left_join(data_hour)
```

Averages per postcode:  

```{r echo=FALSE}
bern_plz_hour %>% 
  st_drop_geometry() %>% 
  ggplot(aes(x = time, y = score, colour = factor(plz))) +
  geom_smooth(se = FALSE) + 
  theme_light()
```

Detailed, individual cell lines by postcode:  

```{r echo=FALSE}
bern_plz_hour %>% 
  st_drop_geometry() %>% 
  ggplot(aes(x = time, y = score, group = tile_id, colour = factor(plz))) +
  geom_line(alpha = 0.5) + 
  facet_wrap(vars(plz), scales = "free_y") + 
  theme_light()
```

Spatial distribution comparing 4AM and 3PM:  

```{r echo=FALSE}
bern_plz_hour %>% 
  filter(hour(time) %in% c(4, 15)) %>% 
  mutate(daytime = ifelse(hour(time) == 4, "Night", "Day")) %>% 
  tm_shape() + 
  tm_polygons(col = "score", alpha = 0.5, palette = "-plasma",
              n = 7, style = "jenks",
              border.col = "white", border.alpha = 0.5) +
  tm_facets(by = "daytime")
```

<!-- ----------------------------------------------------- -->

# Environment

```{r echo=FALSE, results='asis'}
report::report(sessionInfo())
```
